+++ scripts/__init__.py
@@ -1,0 +1,6 @@
+import os
+import sys
+
+parent_dir = os.path.abspath(os.path.join(os.getcwd(), ".."))
+folder_a_path = os.path.join(parent_dir, "scripts")
+sys.path.append(folder_a_path)
